+++ tweaks/panel/src/SCIPanelRoot.m
@@ -16,7 +16,7 @@
 #import <objc/message.h>
 #import <objc/runtime.h>
 
-NSString *SCIVersionString = @"v0.9.23";  // AlbrhiPanel
+NSString *SCIVersionString = @"v0.9.24";  // AlbrhiPanel
 
 ///
 /// Albrhi's own control panel, in the iOS Settings app.